Made accounts modifiable

+++ CoinControl/Services/CategoryService.swift
@@ -7,10 +7,10 @@ import CoreData
 import Foundation
 
 protocol CategoryServiceProtocol {
-    func fetchCategories() throws -> [Category]
-    func fetchCategories(by type: Int16) throws -> [Category]
+    func fetchCategories() throws -> [CategoryModel]
+    func fetchCategories(by type: Int16) throws -> [CategoryModel]
     func addCategory(name: String, icon: String, type: Int16) throws
-    func deleteCategory(_ category: Category) throws
+    func deleteCategory(_ category: CategoryModel) throws
 }
 
 class CategoryService: CategoryServiceProtocol {
@@ -20,17 +20,19 @@ class CategoryService: CategoryServiceProtocol {
         self.context = context
     }
 
-    func fetchCategories() throws -> [Category] {
+    func fetchCategories() throws -> [CategoryModel] {
         let request = Category.fetchRequest()
         request.sortDescriptors = [NSSortDescriptor(keyPath: \Category.name, ascending: true)]
-        return try context.fetch(request)
+        let categories = try context.fetch(request)
+        return categories.map { $0.toModel }
     }
 
-    func fetchCategories(by type: Int16) throws -> [Category] {
+    func fetchCategories(by type: Int16) throws -> [CategoryModel] {
         let request = Category.fetchRequest()
         request.predicate = NSPredicate(format: "type == %d", type)
         request.sortDescriptors = [NSSortDescriptor(keyPath: \Category.name, ascending: true)]
-        return try context.fetch(request)
+        let categories = try context.fetch(request)
+        return categories.map { $0.toModel }
     }
 
     func addCategory(name: String, icon: String, type: Int16) throws {
@@ -43,8 +45,12 @@ class CategoryService: CategoryServiceProtocol {
         try context.save()
     }
 
-    func deleteCategory(_ category: Category) throws {
-        context.delete(category)
-        try context.save()
+    func deleteCategory(_ categoryModel: CategoryModel) throws {
+        let request = Category.fetchRequest()
+        request.predicate = NSPredicate(format: "id == %@", categoryModel.id as CVarArg)
+        if let category = try context.fetch(request).first {
+            context.delete(category)
+            try context.save()
+        }
     }
 }
